cockroach: pass --insecure and --host=localhost when launching service
We've always started CockroachDB insecurely; see #6 for a discussion of
why this is reasonable. As of cockroachdb/cockroach#14703, however, we
must explicitly pass --insecure; running `cockroach start` without
either a cert directory or an explicit --insecure opt-in is an error.
Additionally, we must explicitly specifying `--host=localhost`, as now,
by default, the server listens on all interfaces.