Simplify version bump process
Previously we had to do some manual work to update the formula -
download the file to calculate the SHA256 checksum, update various
fields to reflect the changes. This has been error prone and not
automation friendly.

This patch add a simple helper which uses a template to generate the
brew formula.