Bazelize bump
Previously, running the `generate` make target required Go
pre-installed. It is not the case in our Bazel CI images.

This PR adds Bazel support and builds the release version bumper binary
using Bazel before executing it. Also it explicitly passes the template
file to get rid of some compiled-in magic.